iac/requirements.txt
pulumi>=3.0.0,<4.0.0
pulumi-gcp>=7.0.0,<8.0.0
python-dotenv>=1.0.1,<1.1.0
pulumiverse_time>=0.0.17
pulumi-aws>=6.0.0
pulumi_random>=4.0.0, <5.0.0
pyyaml>=6.0.0,<7.0.0
google-cloud-secret-manager>=2.22.1,<3.0.0
google-api-python-client>=2.159.0, <3.0.0
google-auth>=2.37.0, <3.0.0
pydevd-pycharm~=243.22562.218
cryptography>=44.0.0, <45.0.0